Update archive_logs.py for new log structure, added server setup reference, minute db config.

+++ autotrageur/archive_logs.py
@@ -10,6 +10,7 @@ import os
 import time
 import zipfile
 from datetime import datetime
+from os.path import isdir, join
 
 import schedule
 from docopt import docopt
@@ -20,25 +21,34 @@ from autotrageur.version import VERSION
 def archive_logs():
     """Compresses older log files into zip file in the same directory.
 
+    The bot creates log files in the following format:
+    ./logs/<log-type>/<run-config-id>/<start-timestamp>/<log-files>
+
     Each directory in the 'logs' directory will be searched for
-    concurrent operating bot support. This method will also delete the
-    old file after archiving.
+    concurrent operating bot support. Archives are only created if there
+    logs to archive. Old files are deleted after archiving.
     """
     logging.info('Archive start...')
 
-    for log_dir in filter(dir_filter, os.listdir('logs')):
-        path = 'logs/{}'.format(log_dir)
-        archive_files = filter(lambda x: '.log.' in x, os.listdir(path))
-        zip_file_name = '{}/{}.zip'.format(
-            path,
-            str(datetime.now())
-                .replace(' ', '_').replace('.', '_').replace(':', '_'))
-        zip_file = zipfile.ZipFile(
-            zip_file_name, mode='w', compression=zipfile.ZIP_DEFLATED)
-        for f in archive_files:
-            log_file = '{}/{}'.format(path, f)
-            zip_file.write(log_file)
-            os.remove(log_file)
+    log_types = [d for d in os.listdir('logs') if isdir(join('logs', d))]
+    for log_type in log_types:
+        runs = [d for d in os.listdir(log_type) if isdir(join(log_type, d))]
+        for run in runs:
+            starts = [d for d in os.listdir(run) if isdir(join(run, d))]
+            for start in starts:
+                path = join('logs', log_type, run, start)
+                archive_files = [f for f in os.listdir(path) if '.log.' in f]
+                if archive_files:
+                    zip_file_name = (str(datetime.now())
+                        .replace(' ', '_').replace('.', '_').replace(':', '_'))
+                    zip_file_path = join(path, zip_file_name + '.zip')
+                    zip_file = zipfile.ZipFile(
+                        zip_file_path, mode='w',
+                        compression=zipfile.ZIP_DEFLATED)
+                    for f in archive_files:
+                        log_file = '{}/{}'.format(path, f)
+                        zip_file.write(log_file)
+                        os.remove(log_file)
 
     logging.info('Archive end.')
